Refactor git-tools into lib/scripts/interface three-tier hierarchy
