+++ git-tools/interface/advance_ready.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+"""Promote Todo/Backlog items to Ready when all blockers are closed."""
+import sys
+from pathlib import Path
+
+sys.path.insert(0, str(Path(__file__).parent.parent / "scripts"))
+from project_advance import run
+
+run()
