+++ git-tools/scripts/git-plan.py
@@ -16,20 +16,20 @@ import json
 import os
 import sys
 
-# Allow importing siblings regardless of cwd.
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-import github_client  # noqa: E402
-from issue_utils import (  # noqa: E402
-    get_or_create_issue, append_blocked_by,
-    get_project_node_id, set_item_status,
-    find_active_project,
+LIB = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "lib")
+sys.path.insert(0, LIB)
+import github_client
+from project import (
+    get_project_node_id, set_item_status, find_active_project,
+    create_or_find_project, configure_project_statuses, add_item_to_project,
 )
+from issues import get_or_create_issue, append_blocked_by, load_existing_issues
 
 BLUE  = "\033[34m"
 RESET = "\033[0m"
 
 
-# ── Pre-flight ────────────────────────────────────────────────────────────────
+# ── Pre-flight ───────────────────────────────────────────────────────────────
 
 def check_auth():
     """Verify the token is valid and warn if project scope can't be confirmed."""
@@ -43,175 +43,7 @@ def check_auth():
         pass
 
 
-# ── Project helpers ───────────────────────────────────────────────────────────
-
-def _get_owner_node_id(owner):
-    q = """query($login: String!) { user(login: $login) { id } }"""
-    return github_client.graphql(q, {"login": owner})["data"]["user"]["id"]
-
-
-def create_or_find_project(owner, title, repo):
-    """Create a GitHub Project V2, link it to repo, and return (number, url, node_id).
-
-    Checks for an existing project with the same title first to make re-runs safe.
-    """
-    # Check for existing project
-    list_q = """
-query($login: String!, $first: Int!) {
-  user(login: $login) {
-    projectsV2(first: $first) {
-      nodes { id number title url }
-    }
-  }
-}"""
-    data = github_client.graphql(list_q, {"login": owner, "first": 50})
-    for p in data["data"]["user"]["projectsV2"]["nodes"]:
-        if p["title"] == title:
-            print(f"  (project already exists — reusing #{p['number']})")
-            return p["number"], p["url"], p["id"]
-
-    # Create new project
-    owner_id = _get_owner_node_id(owner)
-    create_m = """
-mutation($ownerId: ID!, $title: String!) {
-  createProjectV2(input: {ownerId: $ownerId, title: $title}) {
-    projectV2 { id number url }
-  }
-}"""
-    project = github_client.graphql(create_m, {"ownerId": owner_id, "title": title})
-    project = project["data"]["createProjectV2"]["projectV2"]
-
-    # Link to repo
-    repo_data   = github_client.rest("GET", f"/repos/{repo}")
-    repo_nid    = repo_data["node_id"]
-    link_m = """
-mutation($projectId: ID!, $repositoryId: ID!) {
-  linkProjectV2ToRepository(input: {projectId: $projectId, repositoryId: $repositoryId}) {
-    repository { id }
-  }
-}"""
-    github_client.graphql(link_m, {"projectId": project["id"], "repositoryId": repo_nid})
-
-    return project["number"], project["url"], project["id"]
-
-
-# ── Status field helpers ──────────────────────────────────────────────────────
-
-EXTRA_STATUSES = [
-    {"name": "Backlog", "color": "GRAY",   "description": "Work that can be done but carries the least priority"},
-    {"name": "Verify",  "color": "ORANGE", "description": "Code is delivered but requires validation to close"},
-    {"name": "Ready",   "color": "BLUE",   "description": "Work that is not pending other tasks to be started"},
-]
-
-
-def configure_project_statuses(project_number, owner):
-    """Add Backlog, Verify, and Ready to the project's Status field.
-
-    Returns (field_id, {name: option_id}) for use when assigning status to new items,
-    or (None, {}) if the Status field is not found.
-    """
-    query = """
-query($login: String!, $number: Int!) {
-  user(login: $login) {
-    projectV2(number: $number) {
-      fields(first: 20) {
-        nodes {
-          ... on ProjectV2SingleSelectField {
-            id name
-            options { id name color description }
-          }
-        }
-      }
-    }
-  }
-}"""
-    data  = github_client.graphql(query, {"login": owner, "number": project_number})
-    nodes = data["data"]["user"]["projectV2"]["fields"]["nodes"]
-    status = next((n for n in nodes if n.get("name") == "Status"), None)
-    if not status:
-        print("  WARNING: Status field not found — skipping")
-        return None, {}
-
-    field_id       = status["id"]
-    existing_names = {o["name"] for o in status["options"]}
-    to_add         = [s for s in EXTRA_STATUSES if s["name"] not in existing_names]
-
-    if not to_add:
-        print("  (status options already configured)")
-        return field_id, {o["name"]: o["id"] for o in status["options"]}
-
-    all_options = [
-        {"name": o["name"], "color": o["color"], "description": o.get("description", "")}
-        for o in status["options"]
-    ] + to_add
-
-    mutation = """
-mutation($fieldId: ID!, $options: [ProjectV2SingleSelectFieldOptionInput!]!) {
-  updateProjectV2Field(input: {
-    fieldId: $fieldId
-    singleSelectOptions: $options
-  }) {
-    projectV2Field {
-      ... on ProjectV2SingleSelectField {
-        id
-        options { id name }
-      }
-    }
-  }
-}"""
-    mut_data = github_client.graphql(mutation, {"fieldId": field_id, "options": all_options})
-    print(f"  Added: {', '.join(s['name'] for s in to_add)}")
-
-    updated_opts = mut_data["data"]["updateProjectV2Field"]["projectV2Field"]["options"]
-    return field_id, {o["name"]: o["id"] for o in updated_opts}
-
-
-# ── Issue helpers ─────────────────────────────────────────────────────────────
-
-def load_existing_issues(repo):
-    """Fetch all issues (open and closed) and return a title → {number, url, node_id} dict.
-
-    Includes closed issues so re-runs don't recreate completed work.
-    """
-    result = {}
-    page   = 1
-    while True:
-        issues = github_client.rest("GET", f"/repos/{repo}/issues",
-                                    params={"state": "all", "per_page": 100, "page": page})
-        if not issues:
-            break
-        for issue in issues:
-            if "pull_request" in issue:
-                continue
-            result[issue["title"]] = {
-                "number":  issue["number"],
-                "url":     issue["html_url"],
-                "node_id": issue["node_id"],
-            }
-        if len(issues) < 100:
-            break
-        page += 1
-    return result
-
-
-# ── Project item helpers ──────────────────────────────────────────────────────
-
-def add_item_to_project(project_id, content_node_id):
-    """Add an issue to a project V2. Idempotent — returns item ID whether new or existing."""
-    mutation = """
-mutation($projectId: ID!, $contentId: ID!) {
-  addProjectV2ItemById(input: {projectId: $projectId, contentId: $contentId}) {
-    item { id }
-  }
-}"""
-    data = github_client.graphql(mutation, {
-        "projectId": project_id,
-        "contentId": content_node_id,
-    })
-    return data["data"]["addProjectV2ItemById"]["item"]["id"]
-
-
-# ── Main ──────────────────────────────────────────────────────────────────────
+# ── Main ─────────────────────────────────────────────────────────────────────
 
 def main():
     parser = argparse.ArgumentParser(
