+++ resources/.config
@@ -1,0 +1,5 @@
+{
+  "owner": "AndresI19",
+  "project_number": 5,
+  "repo": "AndresI19/RS-Agent-Planning"
+}
